Posting ads on agent
Posting advertisements form agent monolith.

+++ agent/src/main/java/com/uns/ftn/agent/domain/Vehicle.java
@@ -20,7 +20,7 @@ public class Vehicle {
    private Long id;
 
    @Column(name = "km_traveled", nullable = false)
-   private double kilometersTraveled;
+   private int kilometersTraveled;
 
    @Column(name = "child_seats", nullable = false)
    private int childSeatNumber;
